Changing to friendly 64 bit integer type rather than 32 bit floating point

+++ daemon/Libraries/APM/ISystemEnvironment.cs
@@ -19,9 +19,9 @@ namespace Spectero.daemon.Libraries.APM
         string GetCpuCacheSize();
 
         // Memory
-        double GetPhysicalMemoryUsed();
-        double GetPhysicalMemoryFree();
-        double GetPhysicalMemoryTotal();
+        long GetPhysicalMemoryUsed();
+        long GetPhysicalMemoryFree();
+        long GetPhysicalMemoryTotal();
 
         // Arch
         bool Is64Bits();
